+++ udata/features/transfer/factories.py
@@ -1,15 +1,12 @@
 # -*- coding: utf-8 -*-
-
 from __future__ import unicode_literals
 
 import factory
-from factory.mongoengine import MongoEngineFactory
 
-from udata import models
-from udata.utils import faker
+from .models import Transfer
 
 
-class TransferFactory(MongoEngineFactory):
+class TransferFactory(factory.mongoengine.MongoEngineFactory):
     class Meta:
-        model = models.Transfer
-    comment = factory.LazyAttribute(lambda o: faker.sentence())
+        model = Transfer
+    comment = factory.Faker('sentence')