+++ CHANGELOG.md
@@ -6,6 +6,8 @@
   [#799](https://github.com/opendatateam/udata/pull/799)
 - Some minor tuning on generic references
   [#801](https://github.com/opendatateam/udata/pull/801)
+- Cleanup factories
+  [#808](https://github.com/opendatateam/udata/pull/808)
 
 ## 1.0.3 (2017-02-21)